+++ codedungeon.Tester/__tests__/cd-tester.js
@@ -1,0 +1,3 @@
+/* eslint-disable */
+
+describe('codedungeon.Tester', () => {})