+++ np.Templating/src/Templating.js
@@ -688,3 +688,41 @@ export async function testInvoke(): Promise<void> {
     logError(pluginJson, error)
   }
 }
+
+export async function templateConvertNote(): Promise<void> {
+  if (typeof Editor.type === 'undefined') {
+    await CommandBar.prompt('Conversion Error', 'Please select the Project Note you would like to convert and try again.')
+    return
+  }
+
+  if (Editor.type !== 'Notes') {
+    await CommandBar.prompt('Conversion Error', 'You can only convert Project Notes')
+    return
+  }
+
+  const note = Editor.content || ''
+
+  const result = new FrontmatterModule().convertProjectNoteToFrontmatter(note)
+  switch (result) {
+    case -1:
+      await CommandBar.prompt('Conversion Falied', 'Unable to convert Project Note.')
+      break
+    case -2:
+      await CommandBar.prompt('Conversion Falied', 'Project Note must have Title (starts with # character)')
+      break
+    case -3:
+      await CommandBar.prompt('Conversion Falied', 'Project Note already in Frontmatter Format')
+      break
+  }
+
+  if (typeof result === 'string') {
+    // select all the text, it will be overwritten by insert of new note
+    Editor.selectAll()
+
+    // replace selected text with converted template
+    Editor.insertTextAtCursor(result.toString())
+
+    // set cursor at the top of the note
+    Editor.highlightByIndex(0, 0)
+  }
+}