+++ np.Templating/src/index.js
@@ -21,6 +21,7 @@ export { templateInvoke } from './Templating'
 export { templateNew } from './Templating'
 export { templateMeetingNote } from './Templating'
 export { templateQuickNote } from './Templating'
+export { templateConvertNote } from './Templating'
 export { templateSamples } from './Templating'
 
 // np.Templating Utility Commands